feat: use config via api on UI
